backend/requirements.txt
asgiref==3.5.0
cffi==1.15.0
chardet==4.0.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.28
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
fonttools==4.31.2
jellyfish==0.9.0
joblib==1.1.0
kiwisolver==1.4.0
matplotlib==3.5.1
networkx==2.7.1
nltk==3.7
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
pdf2image==1.16.0
pdfminer.six==20211012
Pillow==9.0.1
poppler-utils==0.1.0
pycparser==2.21
pyparsing==3.0.7
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2021.3
regex==2022.3.15
scipy==1.8.0
segtok==1.5.11
six==1.16.0
smart-open==5.2.1
sqlparse==0.4.2
tabulate==0.8.9
tqdm==4.63.0
tzdata==2021.5
yake==0.4.8
